+++ src/mainCode/Canon.java
@@ -4,9 +4,9 @@ import java.awt.Color;
 import java.awt.Graphics;
 
 public class Canon {
-	private int x;
-	private int y;
-	private int angle;
+	private double x;
+	private double y;
+	private double angle;
 	
 	public Canon(){
 		this.x = 10;
@@ -14,7 +14,7 @@ public class Canon {
 		this.angle = 45;
 	}
 	
-	public Canon(int x, int y, int angle){
+	public Canon(int x, double y, double angle){
 		this.x = x;
 		this.y = y;
 		this.angle = angle;
@@ -22,13 +22,13 @@ public class Canon {
 	
 	public void draw(Graphics g) {
 		g.setColor(Color.DARK_GRAY);
-		g.drawString("Canon", x, y);
+		g.drawString("Canon", (int) x, (int) y);
 		drawAngle(g);
 	}
 	
 	public void drawAngle(Graphics g) {
 		g.setColor(Color.BLUE);
-		g.drawString("Draw Angle arc", x+20, y+20);
+		g.drawString("Draw Angle arc", (int) x+20, (int) y+20);
 	}
